name

+++ function_pointers/1-array_iterator.c
@@ -10,10 +10,10 @@
  */
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-    unsigned int i;
+    size_t i;
 
     if (array == NULL || action == NULL)
         return;
     for (i = 0; i <= size; i++)
-    action(*(array + i ));
+        action(*(array + i));
 }
